backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.5.1
numpy<2.0.0
PyMuPDF==1.24.9
python-docx==1.1.2
python-pptx==0.6.23
groq==0.9.0
sse-starlette<3.0.0
